Move ModelProvider to a widgets folder
